feat(api): Added prisma seeder
