+++ apps/api/package.json
@@ -5,7 +5,7 @@
   "license": "MIT",
   "prisma": {
     "schema": "src/prisma/schema.prisma",
-    "seed": "ts-node src/prisma/seed.ts"
+    "seed": "ts-node src/prisma/seed/"
   },
   "scripts": {
     "build": "npx tsc && tsc-alias -p tsconfig.json",
